refactor: extract alt-text review from usePhotoImportPanel
266 lines and 18 functions covered four flows: picker, alt-text review,
cropper handoff, and preview/delete.

Extract the alt-text review step into useAltTextReview, and move
fetchSourceFile into photoImportPanel.utils.ts next to the two helpers it
already uses (toImageApiPath, getFileNameFromPath).

The review hook takes an onApproved callback rather than reaching for the
cropper's state, which keeps the handoff one-directional instead of
circular. usePhotoImportPanel: 266 -> 176 lines.

Pure extraction. The hook's returned object has the same 28 keys with the
same names, verified by diffing the key sets, so no consumer changes.

Verified: tsc -p tsconfig.app.json stays at 0 errors; fetchSourceFile
body identical modulo indentation.

Note: photoImportPanel.utils.test.ts cannot currently run — vitest is not
a dependency of lm-client and the npx fallback needs Node 20+ (env is
18.19.1). Typecheck is the only working gate here.
